streamlit-forecasting-app/requirements.txt
streamlit
numpy
pandas
matplotlib.pyplot
datetime
statsmodels
tensorflow